+++ src/commands/Command.cpp
@@ -39,8 +39,6 @@ void Command::fillCommandList() {
     commands.push_back(std::shared_ptr<Command>(new CommandMode()));
     commands.push_back(std::shared_ptr<Command>(new CommandPos()));
     commands.push_back(std::shared_ptr<Command>(new CommandViewmodel()));
-    commands.push_back(std::shared_ptr<Command>(new CommandPigtail()));
-    commands.push_back(std::shared_ptr<Command>(new CommandPonypos()));
     commands.push_back(std::shared_ptr<Command>(new CommandQuit()));
 }